+++ CHANGELOG.md
@@ -19,6 +19,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
       * Layout of the application board is destroyed for those applications which are fully "DONE" but application activation did not take place 
 * User Account
    * Add portal roles
+   * Add Overlay to update users portal roles
 
 ### Change